refactor: restructure AI module by moving controllers and services to dedicated directories; remove obsolete AiQuotaController and AiStreamService files
