+++ apps/api-gateway/src/ai/services/ai-stream.service.spec.ts
